sad old minute
so til that our repo is actually not as big as I had feared
it's a few dozen megs [excluding the archive which I imagine is a similar size] apparently? like really not as ginormous as I had expected tbh. I guess it helps that qxlkbh images are relatively small